tslint fixes and refactor
